Revert "[#13] 도메인별로 묶어 리팩토링하기 (#14)"
This reverts commit b4d81c524c2576ee22f55ea42c5ba8d894f1dc03.
