+++ src/main/java/com/coupang/marketplace/constant/SessionKey.java
@@ -1,0 +1,6 @@
+package com.coupang.marketplace.constant;
+
+public class SessionKey {
+
+    public static final String LOGIN_USER_ID = "USER_ID";
+}
