+++ src/main/java/com/coupang/marketplace/controller/FailResponse.java
@@ -1,0 +1,17 @@
+package com.coupang.marketplace.controller;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+public class FailResponse {
+
+    private StatusEnum status;
+    private String errorMessage;
+
+    @Builder
+    public FailResponse(StatusEnum status, String errorMessage) {
+        this.status = status;
+        this.errorMessage = errorMessage;
+    }
+}
