+++ src/main/java/com/coupang/marketplace/controller/StatusEnum.java
@@ -1,0 +1,24 @@
+package com.coupang.marketplace.controller;
+
+public enum StatusEnum {
+
+    // Success
+    OK(200, "OK"),
+    CREATED(201, "CREATED"),
+    ACCEPTED(202, "ACCEPTED"),
+    NO_CONTENT(204, "NO_CONTENT"),
+    // Fail
+    BAD_REQUEST(400, "BAD_REQUEST"),
+    UNAUTHORIZED(401, "UNAUTHORIZED"),
+    FORBIDDEN(403, "FORBIDDEN"),
+    NOT_FOUND(404, "NOT_FOUND"),
+    INTERNAL_SERER_ERROR(500, "INTERNAL_SERVER_ERROR");
+
+    private int statusCode;
+    private String code;
+
+    StatusEnum(int statusCode, String code) {
+        this.statusCode = statusCode;
+        this.code = code;
+    }
+}
