--- src/main/java/com/coupang/marketplace/global/common/FailResponse.java
@@ -1,17 +1,0 @@
-package com.coupang.marketplace.global.common;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-public class FailResponse {
-
-    private StatusEnum status;
-    private String errorMessage;
-
-    @Builder
-    public FailResponse(StatusEnum status, String errorMessage) {
-        this.status = status;
-        this.errorMessage = errorMessage;
-    }
-}
