--- src/main/java/com/coupang/marketplace/global/common/SuccessResponse.java
@@ -1,19 +1,0 @@
-package com.coupang.marketplace.global.common;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-public class SuccessResponse {
-
-    private StatusEnum status;
-    private String message;
-    private Object data;
-
-    @Builder
-    public SuccessResponse(StatusEnum status, String message, Object data) {
-        this.status = status;
-        this.message = message;
-        this.data = data;
-    }
-}
