--- src/main/java/com/coupang/marketplace/global/constant/SessionKey.java
@@ -1,6 +1,0 @@
-package com.coupang.marketplace.global.constant;
-
-public class SessionKey {
-
-    public static final String LOGIN_USER_ID = "USER_ID";
-}
