--- src/main/java/com/coupang/marketplace/global/util/crypto/CryptoData.java
@@ -1,23 +1,0 @@
-package com.coupang.marketplace.global.util.crypto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-
-@Getter
-public class CryptoData {
-
-    private String plainText;
-    private String salt;
-
-    @Builder
-    public CryptoData(String plainText) {
-        this.plainText = plainText;
-    }
-
-    @Builder(builderClassName= "WithSaltBuilder", builderMethodName = "WithSaltBuilder")
-    public CryptoData(String plainText, String salt) {
-        this.plainText = plainText;
-        this.salt = salt;
-    }
-}
