--- src/main/java/com/coupang/marketplace/global/util/crypto/Encryptor.java
@@ -1,7 +1,0 @@
-package com.coupang.marketplace.global.util.crypto;
-
-public interface Encryptor {
-
-    String encrypt (CryptoData cryptoData);
-
-}
