--- src/main/java/com/coupang/marketplace/user/controller/SignInRequestDto.java
@@ -1,26 +1,0 @@
-package com.coupang.marketplace.user.controller;
-
-import lombok.Builder;
-import lombok.Getter;
-
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-@Getter
-public class SignInRequestDto {
-
-    @NotBlank(message = "이메일을 입력해주세요.")
-    @Email(message = "이메일 양식을 지켜주세요.")
-    private String email;
-
-    @NotBlank(message = "비밀번호를 입력해주세요.")
-    private String password;
-
-    @Builder
-    public SignInRequestDto(@JsonProperty("email") String email, @JsonProperty("password") String password) {
-        this.email = email;
-        this.password = password;
-    }
-}
