--- src/main/java/com/coupang/marketplace/user/repository/UserRepository.java
@@ -1,18 +1,0 @@
-package com.coupang.marketplace.user.repository;
-
-import com.coupang.marketplace.user.domain.User;
-import org.apache.ibatis.annotations.Mapper;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-
-@Mapper
-@Repository
-public interface UserRepository {
-
-    void insertUser(User user);
-
-    Optional<User> findByEmail(String email);
-
-}
