--- src/main/java/com/coupang/marketplace/user/service/LoginService.java
@@ -1,9 +1,0 @@
-package com.coupang.marketplace.user.service;
-
-import com.coupang.marketplace.user.controller.SignInRequestDto;
-
-public interface LoginService {
-
-    void login(SignInRequestDto dto);
-
-}
