--- src/main/java/com/coupang/marketplace/user/service/UserService.java
@@ -1,44 +1,0 @@
-package com.coupang.marketplace.user.service;
-
-import com.coupang.marketplace.user.controller.SignUpRequestDto;
-import com.coupang.marketplace.user.domain.User;
-import com.coupang.marketplace.user.repository.UserRepository;
-import com.coupang.marketplace.global.util.crypto.CryptoData;
-import com.coupang.marketplace.global.util.crypto.Encryptor;
-import com.coupang.marketplace.global.util.crypto.SaltGenerator;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.stereotype.Service;
-
-@RequiredArgsConstructor
-@Service
-public class UserService {
-
-    private final UserRepository userRepository;
-    @Qualifier("sha256Encryptor")
-    private final Encryptor encryptor;
-
-    public void join(SignUpRequestDto dto){
-        if (checkIsUserExist(dto.getEmail())) {
-            throw new IllegalArgumentException("이미 등록된 메일입니다.");
-        }
-
-        String salt = SaltGenerator.generateSalt();
-        CryptoData cryptoData = CryptoData.WithSaltBuilder()
-                .plainText(dto.getPassword())
-                .salt(salt)
-                .build();
-        String encryptedPassword = encryptor.encrypt(cryptoData);
-        User user = dto.toEntity(salt, encryptedPassword);
-
-        userRepository.insertUser(user);
-    }
-
-    public boolean checkIsUserExist (String email) {
-        return userRepository.findByEmail(email).isPresent();
-    }
-
-
-}
-
-
