+++ src/main/java/com/coupang/marketplace/util/Encryptor.java
@@ -1,0 +1,7 @@
+package com.coupang.marketplace.util;
+
+public interface Encryptor {
+
+    String encrypt (CryptoData cryptoData);
+
+}
